+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## new
 
+* [Added] Manual lobuli selection from GUI added
 * [Added] Annotations from ImageJ are readed. The file with annotations is expected to be in the same dir and same name
  with file extension `.roi.zip` 
 * [Fixed] Automatic selection of multiple points in the same lobule